requirements.txt
openai
openai-whisper
sounddevice
soundfile
gTTS
python-dotenv